Struct nvim_rs::exttypes::Buffer

pub struct Buffer<W> where
    W: AsyncWrite + Send + Unpin + 'static,  { /* private fields */ }

A struct representing a neovim buffer. It is specific to a Neovim instance, and calling a method on it will always use this instance.

Implementations

impl<W> Buffer<W> where
W: AsyncWrite + Send + Unpin + 'static,

pub fn new(code_data: Value, neovim: Neovim<W>) -> Buffer<W>

pub fn get_value(&self) -> &Value

Internal value, that represent type

pub async fn line_count(&self) -> Result<i64, Box<CallError>>

since: 1

pub async fn attach(
&self,
send_buffer: bool,
opts: Vec<(Value, Value)>
) -> Result<bool, Box<CallError>>

since: 4

pub async fn detach(&self) -> Result<bool, Box<CallError>>

since: 4

pub async fn get_lines(
&self,
start: i64,
end: i64,
strict_indexing: bool
) -> Result<Vec<String>, Box<CallError>>

since: 1

pub async fn set_lines(
&self,
start: i64,
end: i64,
strict_indexing: bool,
replacement: Vec<String>
) -> Result<(), Box<CallError>>

since: 1

pub async fn set_text(
&self,
start_row: i64,
start_col: i64,
end_row: i64,
end_col: i64,
replacement: Vec<String>
) -> Result<(), Box<CallError>>

since: 7

pub async fn get_offset(&self, index: i64) -> Result<i64, Box<CallError>>

since: 5

pub async fn get_var(&self, name: &str) -> Result<Value, Box<CallError>>

since: 1

pub async fn get_changedtick(&self) -> Result<i64, Box<CallError>>

since: 2

pub async fn get_keymap(
&self,
mode: &str
) -> Result<Vec<Vec<(Value, Value)>>, Box<CallError>>

since: 3

pub async fn set_keymap(
&self,
mode: &str,
lhs: &str,
rhs: &str,
opts: Vec<(Value, Value)>
) -> Result<(), Box<CallError>>

since: 6

pub async fn del_keymap(
&self,
mode: &str,
lhs: &str
) -> Result<(), Box<CallError>>

since: 6

pub async fn get_commands(
&self,
opts: Vec<(Value, Value)>
) -> Result<Vec<(Value, Value)>, Box<CallError>>

since: 4

pub async fn set_var(
&self,
name: &str,
value: Value
) -> Result<(), Box<CallError>>

since: 1

pub async fn del_var(&self, name: &str) -> Result<(), Box<CallError>>

since: 1

pub async fn get_option(&self, name: &str) -> Result<Value, Box<CallError>>

since: 1

pub async fn set_option(
&self,
name: &str,
value: Value
) -> Result<(), Box<CallError>>

since: 1

pub async fn get_name(&self) -> Result<String, Box<CallError>>

since: 1

pub async fn set_name(&self, name: &str) -> Result<(), Box<CallError>>

since: 1

pub async fn is_loaded(&self) -> Result<bool, Box<CallError>>

since: 5

pub async fn delete(
&self,
opts: Vec<(Value, Value)>
) -> Result<(), Box<CallError>>

since: 7

pub async fn is_valid(&self) -> Result<bool, Box<CallError>>

since: 1

pub async fn del_mark(&self, name: &str) -> Result<bool, Box<CallError>>

since: 8

pub async fn set_mark(
&self,
name: &str,
line: i64,
col: i64
) -> Result<bool, Box<CallError>>

since: 8

pub async fn get_mark(&self, name: &str) -> Result<(i64, i64), Box<CallError>>

since: 1

pub async fn get_extmark_by_id(
&self,
ns_id: i64,
id: i64,
opts: Vec<(Value, Value)>
) -> Result<Vec<i64>, Box<CallError>>

since: 7

pub async fn get_extmarks(
&self,
ns_id: i64,
start: Value,
end: Value,
opts: Vec<(Value, Value)>
) -> Result<Vec<Value>, Box<CallError>>

since: 7

pub async fn set_extmark(
&self,
ns_id: i64,
line: i64,
col: i64,
opts: Vec<(Value, Value)>
) -> Result<i64, Box<CallError>>

since: 7

pub async fn del_extmark(
&self,
ns_id: i64,
id: i64
) -> Result<bool, Box<CallError>>

since: 7

pub async fn add_highlight(
&self,
ns_id: i64,
hl_group: &str,
line: i64,
col_start: i64,
col_end: i64
) -> Result<i64, Box<CallError>>

since: 1

pub async fn clear_namespace(
&self,
ns_id: i64,
line_start: i64,
line_end: i64
) -> Result<(), Box<CallError>>

since: 5

pub async fn get_number(&self) -> Result<i64, Box<CallError>>

since: 1

pub async fn clear_highlight(
&self,
ns_id: i64,
line_start: i64,
line_end: i64
) -> Result<(), Box<CallError>>

since: 1

pub async fn set_virtual_text(
&self,
src_id: i64,
line: i64,
chunks: Vec<Value>,
opts: Vec<(Value, Value)>
) -> Result<i64, Box<CallError>>

since: 5