Enum nvim_rs::neovim::Neovim

pub enum Neovim<W> where
    W: AsyncWrite + Send + Unpin + 'static,  {
    Child(Requester<W>, Child),
    Parent(Requester<W>),
    Tcp(Requester<W>),
    UnixSocket(Requester<W>),
}

An active Neovim session.

Variants

Child(Requester<W>, Child)

Parent(Requester<W>)

Tcp(Requester<W>)

UnixSocket(Requester<W>)

Methods

impl<W> Neovim<W> where
W: AsyncWrite + Send + Sync + Unpin + 'static,

pub fn requester(&self) -> Requester<W>

pub async fn call<'_, '_>(
&'_ self,
method: &'_ str,
args: Vec<Value>
) -> Result<Value, Value>

Call can be made only after event loop begin processing

pub async fn ui_attach<'_, '_>(
&'_ mut self,
width: i64,
height: i64,
opts: &'_ UiAttachOptions
) -> Result<(), CallError>

Register as a remote UI.

After this method is called, the client will receive redraw notifications.

pub async fn quit_no_save<'_>(&'_ mut self) -> Result<(), CallError>

Send a quit command to Nvim. The quit command is 'qa!' which will make Nvim quit without saving anything.

impl<W> Neovim<W> where
W: AsyncWrite + Send + Sync + Unpin + 'static,

pub async fn ui_detach<'_>(&'_ self) -> Result<(), CallError>

pub async fn ui_try_resize<'_>(
&'_ self,
width: i64,
height: i64
) -> Result<(), CallError>

pub async fn ui_set_option<'_, '_>(
&'_ self,
name: &'_ str,
value: Value
) -> Result<(), CallError>

pub async fn ui_try_resize_grid<'_>(
&'_ self,
grid: i64,
width: i64,
height: i64
) -> Result<(), CallError>

pub async fn command<'_, '_>(
&'_ self,
command: &'_ str
) -> Result<(), CallError>

pub async fn get_hl_by_name<'_, '_>(
&'_ self,
name: &'_ str,
rgb: bool
) -> Result<Vec<(Value, Value)>, CallError>

pub async fn get_hl_by_id<'_>(
&'_ self,
hl_id: i64,
rgb: bool
) -> Result<Vec<(Value, Value)>, CallError>

pub async fn feedkeys<'_, '_, '_>(
&'_ self,
keys: &'_ str,
mode: &'_ str,
escape_csi: bool
) -> Result<(), CallError>

pub async fn input<'_, '_>(&'_ self, keys: &'_ str) -> Result<i64, CallError>

pub async fn input_mouse<'_, '_, '_, '_>(
&'_ self,
button: &'_ str,
action: &'_ str,
modifier: &'_ str,
grid: i64,
row: i64,
col: i64
) -> Result<(), CallError>

pub async fn replace_termcodes<'_, '_>(
&'_ self,
str: &'_ str,
from_part: bool,
do_lt: bool,
special: bool
) -> Result<String, CallError>

pub async fn command_output<'_, '_>(
&'_ self,
command: &'_ str
) -> Result<String, CallError>

pub async fn eval<'_, '_>(&'_ self, expr: &'_ str) -> Result<Value, CallError>

pub async fn execute_lua<'_, '_>(
&'_ self,
code: &'_ str,
args: Vec<Value>
) -> Result<Value, CallError>

pub async fn call_function<'_, '_>(
&'_ self,
fname: &'_ str,
args: Vec<Value>
) -> Result<Value, CallError>

pub async fn call_dict_function<'_, '_>(
&'_ self,
dict: Value,
fname: &'_ str,
args: Vec<Value>
) -> Result<Value, CallError>

pub async fn strwidth<'_, '_>(&'_ self, text: &'_ str) -> Result<i64, CallError>

pub async fn list_runtime_paths<'_>(&'_ self) -> Result<Vec<String>, CallError>

pub async fn set_current_dir<'_, '_>(
&'_ self,
dir: &'_ str
) -> Result<(), CallError>

pub async fn get_current_line<'_>(&'_ self) -> Result<String, CallError>

pub async fn set_current_line<'_, '_>(
&'_ self,
line: &'_ str
) -> Result<(), CallError>

pub async fn del_current_line<'_>(&'_ self) -> Result<(), CallError>

pub async fn get_var<'_, '_>(
&'_ self,
name: &'_ str
) -> Result<Value, CallError>

pub async fn set_var<'_, '_>(
&'_ self,
name: &'_ str,
value: Value
) -> Result<(), CallError>

pub async fn del_var<'_, '_>(&'_ self, name: &'_ str) -> Result<(), CallError>

pub async fn get_vvar<'_, '_>(
&'_ self,
name: &'_ str
) -> Result<Value, CallError>

pub async fn set_vvar<'_, '_>(
&'_ self,
name: &'_ str,
value: Value
) -> Result<(), CallError>

pub async fn get_option<'_, '_>(
&'_ self,
name: &'_ str
) -> Result<Value, CallError>

pub async fn set_option<'_, '_>(
&'_ self,
name: &'_ str,
value: Value
) -> Result<(), CallError>

pub async fn out_write<'_, '_>(&'_ self, str: &'_ str) -> Result<(), CallError>

pub async fn err_write<'_, '_>(&'_ self, str: &'_ str) -> Result<(), CallError>

pub async fn err_writeln<'_, '_>(
&'_ self,
str: &'_ str
) -> Result<(), CallError>

pub async fn set_current_buf<'_, '_>(
&'_ self,
buffer: &'_ Buffer<W>
) -> Result<(), CallError>

pub async fn set_current_win<'_, '_>(
&'_ self,
window: &'_ Window<W>
) -> Result<(), CallError>

pub async fn set_current_tabpage<'_, '_>(
&'_ self,
tabpage: &'_ Tabpage<W>
) -> Result<(), CallError>

pub async fn create_namespace<'_, '_>(
&'_ self,
name: &'_ str
) -> Result<i64, CallError>

pub async fn get_namespaces<'_>(
&'_ self
) -> Result<Vec<(Value, Value)>, CallError>

pub async fn subscribe<'_, '_>(
&'_ self,
event: &'_ str
) -> Result<(), CallError>

pub async fn unsubscribe<'_, '_>(
&'_ self,
event: &'_ str
) -> Result<(), CallError>

pub async fn get_color_by_name<'_, '_>(
&'_ self,
name: &'_ str
) -> Result<i64, CallError>

pub async fn get_color_map<'_>(
&'_ self
) -> Result<Vec<(Value, Value)>, CallError>

pub async fn get_mode<'_>(&'_ self) -> Result<Vec<(Value, Value)>, CallError>

pub async fn get_keymap<'_, '_>(
&'_ self,
mode: &'_ str
) -> Result<Vec<Vec<(Value, Value)>>, CallError>

pub async fn set_keymap<'_, '_, '_, '_>(
&'_ self,
mode: &'_ str,
lhs: &'_ str,
rhs: &'_ str,
opts: Vec<(Value, Value)>
) -> Result<(), CallError>

pub async fn del_keymap<'_, '_, '_>(
&'_ self,
mode: &'_ str,
lhs: &'_ str
) -> Result<(), CallError>

pub async fn get_commands<'_>(
&'_ self,
opts: Vec<(Value, Value)>
) -> Result<Vec<(Value, Value)>, CallError>

pub async fn get_api_info<'_>(&'_ self) -> Result<Vec<Value>, CallError>

pub async fn set_client_info<'_, '_, '_>(
&'_ self,
name: &'_ str,
version: Vec<(Value, Value)>,
typ: &'_ str,
methods: Vec<(Value, Value)>,
attributes: Vec<(Value, Value)>
) -> Result<(), CallError>

pub async fn get_chan_info<'_>(
&'_ self,
chan: i64
) -> Result<Vec<(Value, Value)>, CallError>

pub async fn list_chans<'_>(&'_ self) -> Result<Vec<Value>, CallError>

pub async fn call_atomic<'_>(
&'_ self,
calls: Vec<Value>
) -> Result<Vec<Value>, CallError>

pub async fn parse_expression<'_, '_, '_>(
&'_ self,
expr: &'_ str,
flags: &'_ str,
highlight: bool
) -> Result<Vec<(Value, Value)>, CallError>

pub async fn list_uis<'_>(&'_ self) -> Result<Vec<Value>, CallError>

pub async fn get_proc_children<'_>(
&'_ self,
pid: i64
) -> Result<Vec<Value>, CallError>

pub async fn get_proc<'_>(&'_ self, pid: i64) -> Result<Value, CallError>

pub async fn select_popupmenu_item<'_>(
&'_ self,
item: i64,
insert: bool,
finish: bool,
opts: Vec<(Value, Value)>
) -> Result<(), CallError>