Struct nvim_rs::runtime::Child

#[must_use = "futures do nothing unless polled"]
pub struct Child { /* fields omitted */ }

Representation of a child process spawned onto an event loop.

This type is also a future which will yield the ExitStatus of the underlying child process. A Child here also provides access to information like the OS-assigned identifier and the stdio streams.

Caveats

Similar to the behavior to the standard library, and unlike the futures paradigm of dropping-implies-cancellation, a spawned process will, by default, continue to execute even after the Child handle has been dropped.

The Command::kill_on_drop method can be used to modify this behavior and kill the child process if the Child wrapper is dropped before it has exited.

Methods

impl Child

pub fn id(&self) -> u32

Returns the OS-assigned process identifier associated with this child.

pub fn kill(&mut self) -> Result<(), Error>

Forces the child to exit.

This is equivalent to sending a SIGKILL on unix platforms.

pub fn stdin(&mut self) -> &mut Option<ChildStdin>

Returns a handle for writing to the child's stdin, if it has been captured.

pub fn stdout(&mut self) -> &mut Option<ChildStdout>

Returns a handle for reading from the child's stdout, if it has been captured.

pub fn stderr(&mut self) -> &mut Option<ChildStderr>

Returns a handle for reading from the child's stderr, if it has been captured.

pub async fn wait_with_output(__arg0: Self) -> Result<Output, Error>

Returns a future that will resolve to an Output, containing the exit status, stdout, and stderr of the child process.

The returned future will simultaneously waits for the child to exit and collect all remaining output on the stdout/stderr handles, returning an Output instance.

The stdin handle to the child process, if any, will be closed before waiting. This helps avoid deadlock: it ensures that the child does not block waiting for input from the parent, while the parent waits for the child to exit.

By default, stdin, stdout and stderr are inherited from the parent. In order to capture the output into this Output it is necessary to create new pipes between parent and child. Use stdout(Stdio::piped()) or stderr(Stdio::piped()), respectively, when creating a Command.