Struct nvim_rs::Window

pub struct Window<W> where
    W: AsyncWrite + Send + Unpin + 'static,  { /* private fields */ }

A struct representing a neovim window. It is specific to a Neovim instance, and calling a method on it will always use this instance.

Implementations

impl<W> Window<W> where
W: AsyncWrite + Send + Unpin + 'static,

pub async fn get_buf(&self) -> Result<Buffer<W>, Box<CallError>>

since: 1

pub async fn get_tabpage(&self) -> Result<Tabpage<W>, Box<CallError>>

since: 1

impl<W> Window<W> where
W: AsyncWrite + Send + Unpin + 'static,

pub fn new(code_data: Value, neovim: Neovim<W>) -> Window<W>

pub fn get_value(&self) -> &Value

Internal value, that represent type

pub async fn set_config(
&self,
config: Vec<(Value, Value)>
) -> Result<(), Box<CallError>>

since: 6

pub async fn get_config(&self) -> Result<Vec<(Value, Value)>, Box<CallError>>

since: 6

pub async fn set_buf(&self, buffer: &Buffer<W>) -> Result<(), Box<CallError>>

since: 5

pub async fn get_cursor(&self) -> Result<(i64, i64), Box<CallError>>

since: 1

pub async fn set_cursor(&self, pos: (i64, i64)) -> Result<(), Box<CallError>>

since: 1

pub async fn get_height(&self) -> Result<i64, Box<CallError>>

since: 1

pub async fn set_height(&self, height: i64) -> Result<(), Box<CallError>>

since: 1

pub async fn get_width(&self) -> Result<i64, Box<CallError>>

since: 1

pub async fn set_width(&self, width: i64) -> Result<(), Box<CallError>>

since: 1

pub async fn get_var(&self, name: &str) -> Result<Value, Box<CallError>>

since: 1

pub async fn set_var(
&self,
name: &str,
value: Value
) -> Result<(), Box<CallError>>

since: 1

pub async fn del_var(&self, name: &str) -> Result<(), Box<CallError>>

since: 1

pub async fn get_option(&self, name: &str) -> Result<Value, Box<CallError>>

since: 1

pub async fn set_option(
&self,
name: &str,
value: Value
) -> Result<(), Box<CallError>>

since: 1

pub async fn get_position(&self) -> Result<(i64, i64), Box<CallError>>

since: 1

pub async fn get_number(&self) -> Result<i64, Box<CallError>>

since: 1

pub async fn is_valid(&self) -> Result<bool, Box<CallError>>

since: 1

pub async fn hide(&self) -> Result<(), Box<CallError>>

since: 7

pub async fn close(&self, force: bool) -> Result<(), Box<CallError>>

since: 6