Trait nvim_rs::rpc::unpack::TryUnpack

pub trait TryUnpack<V> {
    fn try_unpack(self) -> Result<V, Value>;
}

Trait to allow seamless conversion from a Value to the type it contains. In particular, this should never panic.

This is basically a specialized variant of TryInto<V> for Value.

Required methods

fn try_unpack(self) -> Result<V, Value>

Returns the value contained in self.

Errors

Returns Err(self) if self does not contain a value of type V.